+++ src/main/java/com/coursework/workwise/service/ResumeService.java
@@ -3,6 +3,8 @@ package com.coursework.workwise.service;
 import com.coursework.workwise.dto.ResumeCreationDto;
 import com.coursework.workwise.dto.ResumeDto;
 import com.coursework.workwise.entity.Resume;
+import com.coursework.workwise.exception.JobNotFoundException;
+import com.coursework.workwise.exception.ResumeAlreadyExistsException;
 import com.coursework.workwise.exception.ResumeNotFoundException;
 import com.coursework.workwise.mapper.ResumeMapper;
 import com.coursework.workwise.repository.ResumeRepository;
@@ -34,6 +36,10 @@ public class ResumeService {
 
     @Transactional
     public ResumeDto create(ResumeCreationDto resumeCreationDto){
+        boolean userHasResume = resumeRepository.existsByUser(resumeCreationDto.userName());
+        if (userHasResume) {
+            throw new ResumeAlreadyExistsException("User with ID " + resumeCreationDto.userName().getId() + " already has a resume.");
+        }
         return resumeMapper.toDto(resumeRepository.save(resumeMapper.toEntity(resumeCreationDto)));
     }
 
@@ -51,6 +57,9 @@ public class ResumeService {
 
     @Transactional
     public void delete(Long id){
+        if(!resumeRepository.existsById(id)){
+            throw new ResumeNotFoundException("Resume with id " + " not found");
+        }
         resumeRepository.deleteById(id);
     }
 }
